[graphite] More 4-stop gradient tile modes
This splits out the functionality for tiling and colorizing. It passes
t-values around in a manner similar to ganesh in which we actually pass
around a float2 containing t as x and a valid flag (1 or -1) as y, and
if the t is invalid then the colorizer will discard the fragment.

Bug: skia:13302
Change-Id: I901c0f1f6a2c4ffff6b8e3c05d339843d6184086
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/541420
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>