Add a gzip'ed svgDoc to SampleSVG.ttf
The OT-SVG specification allows individual svgDoc inside the SVG table
to be optionaly gzip'ed. Add a glyph to SampleSVG.ttf which uses a
compressed svgDoc so that this can be tested.

This font crashes vanilla FreeType 2.12.0 and 2.12.1. This issue was
fixed with [0] and backported to Debian in 2.12.1+dfsg-3 [1].

[0] https://gitlab.freedesktop.org/freetype/freetype/-/commit/c26872ed59cba3af2f407b5eefc92fcec92aa52b
[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1013094

Change-Id: I45e115a743b8aa4d545f34c9668597d22e0a2bf4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/545779
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Herb Derby <herb@google.com>