Switch to using LeftTop for Transformed and SDFT glyphs
Currently, the position and source rectangle are stored for each
glyph. The atlas entry has the needed information to reconstruct
the rectangle. Switch to storing the LeftTop of each glyph,
and use the rect from the atlas entry to get the width and height.
Adjust all the serialize/deserialize code to use the new format.

Change-Id: Ifb8524fb3a53d2f52059707918c732804e7f093a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/547285
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Herb Derby <herb@google.com>