Remove bit-field optimization for RemoteGlyphCache
I have two reasons for doing this. The first is that I don't think
it was doing anything. In order to set the bit field the statement:

digest->canDrawAsMask() && digest->canDrawAsSDFT()

had to be true. Which I think is impossible because this is
in the mask (not SDFT) case, nd the mask is never SDF when 
preparing for mask drawing, since SDF masks will be prepared 
when preparing for SDF drawing.

Second, this code is optimized for TextBlob use. Optimizing for
Slugs means that we will need to keep more information around.
This bit vector will never be useful for Slugs.

Change-Id: Ic9233a8a061b58a53d773ae278756b0b714f7c73
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/548481
Commit-Queue: Herb Derby <herb@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>