[graphite] ClipStack intersection logic fixes
The previous logic in handling complex bbox intersections in Graphite's
clip stack was fundamentally flawed. It's now replaced with the correct
solution, which is to use the separating axis theorem, instead of just
checking corner points.

The flow through the intersection case has been re-worked to move the
rect-stays-rect dual to the forefront, since that is the most common and
simplest scenario.

Also adds an extra fast check for contains() to avoid more expensive
analysis when the outer bounding boxes already show containment is not
possible.

Found a minor bug in the creation of the clip stack elements where it
was only applying the rect-stays-rect transform to map shapes to device
space when it was *only* rect-stays-rect and not simpler. Importantly,
this meant that translate and positive scale+translates were remaining
un-mapped so comparisons between elements would always then hit the
more expensive "differing coordinate space" codepaths. Previously, the
only cases that would trigger this would be mirrors and 90 degree
rotations.

Change-Id: I328ca1515ff578a617ef18992819c7c9465cb30d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/548478
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>