Assert when creating nullptr spans with a nonzero length.
According to https://en.cppreference.com/w/cpp/container/span/span (2),
it  is undefined behavior if [first, first + count) is not a valid
range, so it is standards-conforming to assert on this span.

This caught one case in GlyphRun where we were creating an invalid
span (which was apparently harmless in practice).

Change-Id: Ia0661b3bd13c2c5773492960eea11b1a294072d5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/550499
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Herb Derby <herb@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>