Fix not using texture barrier on StrokeTessOp.
Previously we were overwriting the renderpassXferBarriers flag on
ProgramInfo to set it to kNone. This flag is meant to say whether or not
the entire render pass uses barriers or not. This is needed in Vulkan
because all pipelines in a render pass that has an input attachment
must bind the input attachment regardless if it is used or not. So the
pipeline must be created with a layout for an input attachment
descriptor set.

This change just removes to performance optimization to only use the
barrier on the stencil and not fill draw. This use case shouldn't
come up too often and also shouldn't be a big perf hit regardless.
The way GrAppliedClip is created/used it is hard for us to create
multiple different Pipeline objects: one for stencil and one for the
fill.

Bug: skia:13402
Change-Id: I15ce74b4d41b90d3dd4169a1f4fb77ed87c8b26d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/549898
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Greg Daniel <egdaniel@google.com>